1 task, ls, pwd, cat, exit
